v.kriging, v.nnstat addons: attempt to fix Makefile; fix compiler warning -Wabsolute-value and return in void function
git-svn-id: https://svn.osgeo.org/grass/grass-addons@66374 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ grass7/vector/v.nnstat/utils.c
@@ -25,7 +25,7 @@ double bearing(double x0, double x1, double y0, double y1)
         return -9999;
     }
 
-    us = atan(fabsf(dy / dx));
+    us = atan(fabs(dy / dx));
     if (dx == 0. && dy > 0.) {
         us = 0.5 * PI;
     }
